Move proof stuff to its own file
