+++ packages/notarization/proofs.go
@@ -1,0 +1,96 @@
+package notarization
+
+import (
+	"github.com/celestiaorg/smt"
+	"github.com/cockroachdb/errors"
+	"github.com/iotaledger/goshimmer/packages/epoch"
+	"github.com/iotaledger/goshimmer/packages/ledger/utxo"
+	"github.com/iotaledger/goshimmer/packages/ledger/vm/devnetvm"
+	"github.com/iotaledger/goshimmer/packages/tangle"
+	"github.com/iotaledger/hive.go/generics/lo"
+	"golang.org/x/crypto/blake2b"
+)
+
+type CommitmentProof struct {
+	EI    epoch.Index
+	proof smt.SparseMerkleProof
+	root  []byte
+}
+
+// GetBlockInclusionProof gets the proof of the inclusion (acceptance) of a block.
+func (m *Manager) GetBlockInclusionProof(blockID tangle.MessageID) (*CommitmentProof, error) {
+	var ei epoch.Index
+	m.tangle.Storage.Message(blockID).Consume(func(block *tangle.Message) {
+		t := block.IssuingTime()
+		ei = m.epochManager.TimeToEI(t)
+	})
+	proof, err := m.epochCommitmentFactory.ProofTangleRoot(ei, blockID)
+	if err != nil {
+		return nil, err
+	}
+	return proof, nil
+}
+
+// GetTransactionInclusionProof gets the proof of the inclusion (acceptance) of a transaction.
+func (m *Manager) GetTransactionInclusionProof(transactionID utxo.TransactionID) (*CommitmentProof, error) {
+	var ei epoch.Index
+	m.tangle.Ledger.Storage.CachedTransaction(transactionID).Consume(func(tx utxo.Transaction) {
+		t := tx.(*devnetvm.Transaction).Essence().Timestamp()
+		ei = m.epochManager.TimeToEI(t)
+	})
+	proof, err := m.epochCommitmentFactory.ProofStateMutationRoot(ei, transactionID)
+	if err != nil {
+		return nil, err
+	}
+	return proof, nil
+}
+
+
+func (f *EpochCommitmentFactory) verifyRoot(proof CommitmentProof, key []byte, value []byte) bool {
+	return smt.VerifyProof(proof.proof, proof.root, key, value, lo.PanicOnErr(blake2b.New256(nil)))
+}
+
+// ProofStateRoot returns the merkle proof for the outputID against the state root.
+func (f *EpochCommitmentFactory) ProofStateRoot(ei epoch.Index, outID utxo.OutputID) (*CommitmentProof, error) {
+	key := outID.Bytes()
+	root := f.commitmentTrees[ei].tangleTree.Root()
+	proof, err := f.stateRootTree.ProveForRoot(key, root)
+	if err != nil {
+		return nil, errors.Wrap(err, "could not generate the state root proof")
+	}
+	return &CommitmentProof{ei, proof, root}, nil
+}
+
+// ProofStateMutationRoot returns the merkle proof for the transactionID against the state mutation root.
+func (f *EpochCommitmentFactory) ProofStateMutationRoot(ei epoch.Index, txID utxo.TransactionID) (*CommitmentProof, error) {
+	key := txID.Bytes()
+	root := f.commitmentTrees[ei].stateMutationTree.Root()
+	proof, err := f.commitmentTrees[ei].stateMutationTree.ProveForRoot(key, root)
+	if err != nil {
+		return nil, errors.Wrap(err, "could not generate the state mutation root proof")
+	}
+	return &CommitmentProof{ei, proof, root}, nil
+}
+
+// ProofTangleRoot returns the merkle proof for the blockID against the tangle root.
+func (f *EpochCommitmentFactory) ProofTangleRoot(ei epoch.Index, blockID tangle.MessageID) (*CommitmentProof, error) {
+	key := blockID.Bytes()
+	root := f.commitmentTrees[ei].tangleTree.Root()
+	proof, err := f.commitmentTrees[ei].tangleTree.ProveForRoot(key, root)
+	if err != nil {
+		return nil, errors.Wrap(err, "could not generate the tangle root proof")
+	}
+	return &CommitmentProof{ei, proof, root}, nil
+}
+
+// VerifyTangleRoot verify the provided merkle proof against the tangle root.
+func (f *EpochCommitmentFactory) VerifyTangleRoot(proof CommitmentProof, blockID tangle.MessageID) bool {
+	key := blockID.Bytes()
+	return f.verifyRoot(proof, key, key)
+}
+
+// VerifyStateMutationRoot verify the provided merkle proof against the state mutation root.
+func (f *EpochCommitmentFactory) VerifyStateMutationRoot(proof CommitmentProof, transactionID utxo.TransactionID) bool {
+	key := transactionID.Bytes()
+	return f.verifyRoot(proof, key, key)
+}
